resrepo: Easy git repo management for reproducible science

A package to manage git repositories used for research that contain a with a mix of code and data. resrepo is based on a well documented template with clear guidelines on how the repository should be structured, what file naming conventions to use, and how to manage large data that can not be stored in git repositories, encouraging reproducible science. In particular, resrepo provides tools to ensure that data are available and up-to-date both locally and in the cloud, as well as simplifying the process of locating directories within the repository from R scripts.

See also

Author

Maintainer: Andrea Manica am351@cam.ac.uk

Other contributors:

Evie Carter [contributor]

Andrea Vittorio Pozzi [contributor]

Margherita Colucci [contributor]

Cecilia Padilla-Iglesias [contributor]

Michela Leonardi [contributor]

Paul Whitelaw [contributor]

Eirlys Tysall [contributor]

Ondrej Pelanek [contributor]

Zhe Xue [contributor]

Anahit Hovhannisyan [contributor]

Dominik Jud [contributor]

Cassandra Gunasekaram [contributor]

Elizabeth Pearmain [contributor]

James Kennerley [contributor]